Refactor the Table component
